Fixed icons for SelectFBDialog in QuickFixOption
Also Refactored FBTreeNodeLabelProvider into it's own class
